requirements.txt
Jpype1-Py3
selenium
